refactor: rename and update references from `dragonfruit-slicer-v3` to `dragonfruit-slicing-engine`
- Updated all references in the codebase to reflect the new module name.
