+++ rust/dragonfruit-islands/Cargo.toml
@@ -8,7 +8,7 @@ name = "dragonfruit_islands"
 path = "src/lib.rs"
 
 [dependencies]
-dragonfruit-slicer-v3 = { path = "../dragonfruit-slicer-v3" }
+dragonfruit-slicing-engine = { path = "../dragonfruit-slicing-engine" }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 rayon = "1.10"
