+++ rust/dragonfruit-islands/src/geometry.rs
@@ -1,3 +1,3 @@
 //! Re-exports geometry types from the slicer crate.
 
-pub use dragonfruit_slicer_v3::geometry::{parse_triangles, Triangle, Vec3};
+pub use dragonfruit_slicing_engine::geometry::{parse_triangles, Triangle, Vec3};
