--- rust/dragonfruit-slicer-v3/src/engine.rs
@@ -1,640 +1,0 @@
-//! V3 engine orchestration and validation layer.
-
-use crate::encoders::registry::{find_encoder, supported_output_formats};
-use crate::geometry::{parse_triangles, project_triangles_inplace};
-use crate::index::build_layer_index;
-use crate::metrics::SlicingPerfV3;
-use crate::pipeline::{render_layers_bounded, render_layers_rle, render_layers_rle_encoded};
-use crate::types::{
-    LayerAreaStatsV3, ProgressCallbackV3, RenderedLayersV3, SliceArtifactV3, SliceJobV3,
-    SliceProgressPhaseV3, SliceProgressUpdateV3,
-};
-use std::path::Path;
-use std::sync::atomic::AtomicBool;
-use std::sync::Arc;
-use thiserror::Error;
-
-#[derive(Debug, Error)]
-pub enum SlicerV3Error {
-    #[error("cancelled")]
-    Cancelled,
-    #[error("unsupported output format: {0}")]
-    UnsupportedOutput(String),
-    #[error("invalid dimensions {width}x{height}")]
-    InvalidDimensions { width: u32, height: u32 },
-    #[error(
-        "invalid layer settings: layer_height_mm={layer_height_mm}, total_layers={total_layers}"
-    )]
-    InvalidLayerSettings {
-        layer_height_mm: f32,
-        total_layers: u32,
-    },
-    #[error("invalid build volume dimensions: build_width_mm={build_width_mm}, build_depth_mm={build_depth_mm}")]
-    InvalidBuildVolume {
-        build_width_mm: f32,
-        build_depth_mm: f32,
-    },
-    #[error("invalid triangle buffer length: expected multiple of 9, got {0}")]
-    InvalidTriangleBuffer(usize),
-    #[error("png encode failed: {0}")]
-    Png(String),
-    #[error("zip encode failed: {0}")]
-    Zip(String),
-    #[error("json encode failed: {0}")]
-    Json(String),
-    #[error("missing rendered layer payload: {0}")]
-    MissingRenderedLayerPayload(String),
-}
-
-fn validate_job(job: &SliceJobV3) -> Result<(), SlicerV3Error> {
-    if job.width_px == 0
-        || job.height_px == 0
-        || job.source_width_px == 0
-        || job.source_height_px == 0
-    {
-        return Err(SlicerV3Error::InvalidDimensions {
-            width: job.width_px,
-            height: job.height_px,
-        });
-    }
-    if !(job.layer_height_mm.is_finite() && job.layer_height_mm > 0.0) || job.total_layers == 0 {
-        return Err(SlicerV3Error::InvalidLayerSettings {
-            layer_height_mm: job.layer_height_mm,
-            total_layers: job.total_layers,
-        });
-    }
-    if !(job.build_width_mm.is_finite() && job.build_width_mm > 0.0)
-        || !(job.build_depth_mm.is_finite() && job.build_depth_mm > 0.0)
-    {
-        return Err(SlicerV3Error::InvalidBuildVolume {
-            build_width_mm: job.build_width_mm,
-            build_depth_mm: job.build_depth_mm,
-        });
-    }
-    if job.triangles_xyz.len() % 9 != 0 {
-        return Err(SlicerV3Error::InvalidTriangleBuffer(
-            job.triangles_xyz.len(),
-        ));
-    }
-    Ok(())
-}
-
-/// Clean-room V3 entry point with full pipeline:
-/// parse triangles -> build layer index -> bounded parallel render -> zip archive encode.
-pub fn slice_with_progress_v3(
-    job: &SliceJobV3,
-    on_progress: Option<ProgressCallbackV3>,
-    cancel_flag: Option<&AtomicBool>,
-) -> Result<SliceArtifactV3, SlicerV3Error> {
-    let Some(encoder) = find_encoder(&job.output_format) else {
-        return Err(SlicerV3Error::UnsupportedOutput(format!(
-            "{} (supported: {})",
-            job.output_format,
-            supported_output_formats().join(", ")
-        )));
-    };
-    let requires_area_stats = encoder.requires_area_stats();
-    let requires_png_layers = encoder.requires_png_layers();
-    let requires_raw_mask_layers = encoder.requires_raw_mask_layers();
-
-    // RLE path: no full-image pixel buffer — fastest for formats like CTBv5.
-    if !requires_png_layers {
-        if let Some(mut rle_enc) = encoder.create_rle_stream_encoder(job)? {
-            let total_start = std::time::Instant::now();
-            let job_total_layers = job.total_layers;
-            let progress_total = job_total_layers.saturating_add(1);
-
-            let slicing_progress = on_progress.as_ref().map(|cb| {
-                let cb = cb.clone();
-                Arc::new(move |update: SliceProgressUpdateV3| {
-                    cb(SliceProgressUpdateV3 {
-                        done: update.done.min(job_total_layers),
-                        total: progress_total,
-                        phase: SliceProgressPhaseV3::Slicing,
-                    });
-                }) as ProgressCallbackV3
-            });
-
-            // Parallel-encode path: rasterize + encode PNG in rayon workers.
-            let (_rendered_layers, layer_area_stats, mut perf) = if let Some(encode_fn) =
-                rle_enc.parallel_encode_fn()
-            {
-                let mut store_sink = |idx: u32, bytes: Vec<u8>| -> Result<(), SlicerV3Error> {
-                    rle_enc.store_encoded_layer(idx, bytes);
-                    Ok(())
-                };
-                slice_and_rasterize_rle_encoded_v3(
-                    job,
-                    encode_fn,
-                    &mut store_sink,
-                    slicing_progress,
-                    cancel_flag,
-                )?
-            } else {
-                let mut rle_sink =
-                    |idx: u32, runs: Vec<crate::rle::RleRun>| rle_enc.consume_rle_layer(idx, runs);
-                slice_and_rasterize_rle_v3(job, &mut rle_sink, slicing_progress, cancel_flag)?
-            };
-
-            rle_enc.set_area_stats(layer_area_stats);
-
-            if let Some(cb) = on_progress.as_ref() {
-                cb(SliceProgressUpdateV3 {
-                    done: job_total_layers,
-                    total: progress_total,
-                    phase: SliceProgressPhaseV3::Finalizing,
-                });
-            }
-
-            let encode_start = std::time::Instant::now();
-            let bytes = rle_enc.finalize_to_bytes()?;
-
-            if let Some(cb) = on_progress.as_ref() {
-                cb(SliceProgressUpdateV3 {
-                    done: progress_total,
-                    total: progress_total,
-                    phase: SliceProgressPhaseV3::Finalizing,
-                });
-            }
-
-            perf.archive_encode_ns = encode_start.elapsed().as_nanos() as u64;
-            perf.total_ns = total_start.elapsed().as_nanos() as u64;
-            perf.layers = job.total_layers;
-
-            return Ok(SliceArtifactV3 { bytes, perf });
-        }
-    }
-
-    if !requires_png_layers && requires_raw_mask_layers {
-        if let Some(mut stream_encoder) = encoder.create_raw_mask_stream_encoder(job)? {
-            let total_start = std::time::Instant::now();
-            let job_total_layers = job.total_layers;
-            let progress_total = job_total_layers.saturating_add(1);
-            let mut raw_mask_sink = |layer_index: u32, raw_mask: Vec<u8>| {
-                stream_encoder.consume_raw_mask_layer(layer_index, raw_mask)
-            };
-
-            let slicing_progress = on_progress.as_ref().map(|cb| {
-                let cb = cb.clone();
-                Arc::new(move |update: SliceProgressUpdateV3| {
-                    cb(SliceProgressUpdateV3 {
-                        done: update.done.min(job_total_layers),
-                        total: progress_total,
-                        phase: SliceProgressPhaseV3::Slicing,
-                    });
-                }) as ProgressCallbackV3
-            });
-
-            let (_rendered_layers, _layer_area_stats, mut perf) = slice_and_rasterize_v3(
-                job,
-                false,
-                requires_png_layers,
-                false,
-                Some(&mut raw_mask_sink),
-                slicing_progress,
-                cancel_flag,
-            )?;
-
-            if let Some(cb) = on_progress.as_ref() {
-                cb(SliceProgressUpdateV3 {
-                    done: job_total_layers,
-                    total: progress_total,
-                    phase: SliceProgressPhaseV3::Finalizing,
-                });
-            }
-
-            let encode_start = std::time::Instant::now();
-            let bytes = stream_encoder.finalize_to_bytes()?;
-
-            if let Some(cb) = on_progress.as_ref() {
-                cb(SliceProgressUpdateV3 {
-                    done: progress_total,
-                    total: progress_total,
-                    phase: SliceProgressPhaseV3::Finalizing,
-                });
-            }
-
-            perf.archive_encode_ns = encode_start.elapsed().as_nanos() as u64;
-            perf.total_ns = total_start.elapsed().as_nanos() as u64;
-            perf.layers = job.total_layers;
-
-            return Ok(SliceArtifactV3 { bytes, perf });
-        }
-    }
-
-    let total_start = std::time::Instant::now();
-    let (rendered_layers, layer_area_stats, mut perf) = slice_and_rasterize_v3(
-        job,
-        requires_area_stats,
-        requires_png_layers,
-        requires_raw_mask_layers,
-        None,
-        on_progress.clone(),
-        cancel_flag,
-    )?;
-
-    let encode_units = encoder
-        .estimate_encode_progress_units(&rendered_layers)
-        .max(1);
-    let progress_total = job.total_layers.saturating_add(encode_units);
-
-    let encode_progress = on_progress.as_ref().map(|cb| {
-        let cb = cb.clone();
-        move |done: u32, total: u32| {
-            let safe_total = total.max(1);
-            let clamped_done = done.min(safe_total);
-            let normalized =
-                ((clamped_done as u64) * (encode_units as u64) / (safe_total as u64)) as u32;
-            cb(SliceProgressUpdateV3 {
-                done: job.total_layers.saturating_add(normalized),
-                total: progress_total,
-                phase: SliceProgressPhaseV3::Encoding,
-            });
-        }
-    });
-
-    let encode_start = std::time::Instant::now();
-    let bytes = dispatch_encode_by_format(
-        job,
-        &rendered_layers,
-        &layer_area_stats,
-        encode_progress.as_ref().map(|cb| cb as &dyn Fn(u32, u32)),
-    )?;
-
-    if let Some(cb) = on_progress.as_ref() {
-        cb(SliceProgressUpdateV3 {
-            done: progress_total,
-            total: progress_total,
-            phase: SliceProgressPhaseV3::Finalizing,
-        });
-    }
-
-    perf.archive_encode_ns = encode_start.elapsed().as_nanos() as u64;
-    perf.total_ns = total_start.elapsed().as_nanos() as u64;
-    perf.layers = job.total_layers;
-
-    Ok(SliceArtifactV3 { bytes, perf })
-}
-
-/// Fast raster stage that produces RLE runs per layer — no pixel buffers.
-pub fn slice_and_rasterize_rle_v3(
-    job: &SliceJobV3,
-    on_rle_layer: impl FnMut(u32, Vec<crate::rle::RleRun>) -> Result<(), SlicerV3Error>,
-    on_progress: Option<ProgressCallbackV3>,
-    cancel_flag: Option<&AtomicBool>,
-) -> Result<(RenderedLayersV3, Vec<LayerAreaStatsV3>, SlicingPerfV3), SlicerV3Error> {
-    validate_job(job)?;
-
-    let mut triangles = parse_triangles(&job.triangles_xyz);
-    project_triangles_inplace(&mut triangles, job);
-    let index_start = std::time::Instant::now();
-    let layer_index = build_layer_index(&triangles, job.total_layers, job.layer_height_mm);
-    let index_ns = index_start.elapsed().as_nanos() as u64;
-
-    let (rendered_layers, layer_area_stats, mut perf) = render_layers_rle(
-        job,
-        &triangles,
-        &layer_index,
-        on_rle_layer,
-        on_progress,
-        cancel_flag,
-    )?;
-    perf.index_build_ns = index_ns;
-
-    Ok((rendered_layers, layer_area_stats, perf))
-}
-
-/// Fast raster stage that encodes RLE runs into layer bytes in parallel rayon workers.
-pub fn slice_and_rasterize_rle_encoded_v3(
-    job: &SliceJobV3,
-    encode_fn: Arc<
-        dyn Fn(u32, &[crate::rle::RleRun]) -> Result<Vec<u8>, SlicerV3Error> + Send + Sync,
-    >,
-    on_encoded_layer: impl FnMut(u32, Vec<u8>) -> Result<(), SlicerV3Error>,
-    on_progress: Option<ProgressCallbackV3>,
-    cancel_flag: Option<&AtomicBool>,
-) -> Result<(RenderedLayersV3, Vec<LayerAreaStatsV3>, SlicingPerfV3), SlicerV3Error> {
-    validate_job(job)?;
-
-    let mut triangles = parse_triangles(&job.triangles_xyz);
-    project_triangles_inplace(&mut triangles, job);
-    let index_start = std::time::Instant::now();
-    let layer_index = build_layer_index(&triangles, job.total_layers, job.layer_height_mm);
-    let index_ns = index_start.elapsed().as_nanos() as u64;
-
-    let (rendered_layers, layer_area_stats, mut perf) = render_layers_rle_encoded(
-        job,
-        &triangles,
-        &layer_index,
-        encode_fn,
-        on_encoded_layer,
-        on_progress,
-        cancel_flag,
-    )?;
-    perf.index_build_ns = index_ns;
-
-    Ok((rendered_layers, layer_area_stats, perf))
-}
-
-/// Format-agnostic geometry/index/raster stage that outputs layer PNG bytes.
-pub fn slice_and_rasterize_v3(
-    job: &SliceJobV3,
-    requires_area_stats: bool,
-    emit_png_layers: bool,
-    emit_raw_mask_layers: bool,
-    on_raw_mask_layer: Option<&mut dyn FnMut(u32, Vec<u8>) -> Result<(), SlicerV3Error>>,
-    on_progress: Option<ProgressCallbackV3>,
-    cancel_flag: Option<&AtomicBool>,
-) -> Result<(RenderedLayersV3, Vec<LayerAreaStatsV3>, SlicingPerfV3), SlicerV3Error> {
-    validate_job(job)?;
-
-    let mut triangles = parse_triangles(&job.triangles_xyz);
-    project_triangles_inplace(&mut triangles, job);
-    let index_start = std::time::Instant::now();
-    let layer_index = build_layer_index(&triangles, job.total_layers, job.layer_height_mm);
-    let index_ns = index_start.elapsed().as_nanos() as u64;
-
-    let (rendered_layers, layer_area_stats, mut perf) = render_layers_bounded(
-        job,
-        &triangles,
-        &layer_index,
-        requires_area_stats,
-        emit_png_layers,
-        emit_raw_mask_layers,
-        on_raw_mask_layer,
-        on_progress,
-        cancel_flag,
-    )?;
-    perf.index_build_ns = index_ns;
-
-    Ok((rendered_layers, layer_area_stats, perf))
-}
-
-/// Encode rendered layers through a registered format encoder.
-pub fn dispatch_encode_by_format(
-    job: &SliceJobV3,
-    rendered_layers: &RenderedLayersV3,
-    layer_area_stats: &[LayerAreaStatsV3],
-    on_encode_progress: Option<&dyn Fn(u32, u32)>,
-) -> Result<Vec<u8>, SlicerV3Error> {
-    let Some(encoder) = find_encoder(&job.output_format) else {
-        return Err(SlicerV3Error::UnsupportedOutput(format!(
-            "{} (supported: {})",
-            job.output_format,
-            supported_output_formats().join(", ")
-        )));
-    };
-    encoder.encode_container_from_rendered_layers_with_progress(
-        job,
-        rendered_layers,
-        layer_area_stats,
-        on_encode_progress,
-    )
-}
-
-/// Encode rendered layers through a registered format encoder directly to disk.
-pub fn dispatch_encode_by_format_to_path(
-    job: &SliceJobV3,
-    rendered_layers: &RenderedLayersV3,
-    layer_area_stats: &[LayerAreaStatsV3],
-    output_path: &Path,
-    on_encode_progress: Option<&dyn Fn(u32, u32)>,
-) -> Result<(), SlicerV3Error> {
-    let Some(encoder) = find_encoder(&job.output_format) else {
-        return Err(SlicerV3Error::UnsupportedOutput(format!(
-            "{} (supported: {})",
-            job.output_format,
-            supported_output_formats().join(", ")
-        )));
-    };
-    encoder.encode_container_to_path_with_progress(
-        job,
-        rendered_layers,
-        layer_area_stats,
-        output_path,
-        on_encode_progress,
-    )
-}
-
-/// Path-oriented entrypoint that writes final archive bytes directly to disk.
-///
-/// This avoids materializing the final encoded artifact in-memory at the call
-/// boundary, reducing bridge and copy overhead in desktop/Tauri flows.
-pub fn slice_with_progress_v3_to_path(
-    job: &SliceJobV3,
-    output_path: &Path,
-    on_progress: Option<ProgressCallbackV3>,
-    cancel_flag: Option<&AtomicBool>,
-) -> Result<SlicingPerfV3, SlicerV3Error> {
-    let Some(encoder) = find_encoder(&job.output_format) else {
-        return Err(SlicerV3Error::UnsupportedOutput(format!(
-            "{} (supported: {})",
-            job.output_format,
-            supported_output_formats().join(", ")
-        )));
-    };
-    let requires_area_stats = encoder.requires_area_stats();
-    let requires_png_layers = encoder.requires_png_layers();
-    let requires_raw_mask_layers = encoder.requires_raw_mask_layers();
-
-    // RLE path: no full-image pixel buffer — fastest for formats like CTBv5.
-    if !requires_png_layers {
-        if let Some(mut rle_enc) = encoder.create_rle_stream_encoder(job)? {
-            let total_start = std::time::Instant::now();
-            let job_total_layers = job.total_layers;
-            let progress_total = job_total_layers.saturating_add(1);
-
-            let slicing_progress = on_progress.as_ref().map(|cb| {
-                let cb = cb.clone();
-                Arc::new(move |update: SliceProgressUpdateV3| {
-                    cb(SliceProgressUpdateV3 {
-                        done: update.done.min(job_total_layers),
-                        total: progress_total,
-                        phase: SliceProgressPhaseV3::Slicing,
-                    });
-                }) as ProgressCallbackV3
-            });
-
-            let (_rendered_layers, layer_area_stats, mut perf) = if let Some(encode_fn) =
-                rle_enc.parallel_encode_fn()
-            {
-                let mut store_sink = |idx: u32, bytes: Vec<u8>| -> Result<(), SlicerV3Error> {
-                    rle_enc.store_encoded_layer(idx, bytes);
-                    Ok(())
-                };
-                slice_and_rasterize_rle_encoded_v3(
-                    job,
-                    encode_fn,
-                    &mut store_sink,
-                    slicing_progress,
-                    cancel_flag,
-                )?
-            } else {
-                let mut rle_sink =
-                    |idx: u32, runs: Vec<crate::rle::RleRun>| rle_enc.consume_rle_layer(idx, runs);
-                slice_and_rasterize_rle_v3(job, &mut rle_sink, slicing_progress, cancel_flag)?
-            };
-
-            rle_enc.set_area_stats(layer_area_stats);
-
-            if let Some(cb) = on_progress.as_ref() {
-                cb(SliceProgressUpdateV3 {
-                    done: job_total_layers,
-                    total: progress_total,
-                    phase: SliceProgressPhaseV3::Finalizing,
-                });
-            }
-
-            let encode_start = std::time::Instant::now();
-            rle_enc.finalize_to_path(output_path)?;
-
-            if let Some(cb) = on_progress.as_ref() {
-                cb(SliceProgressUpdateV3 {
-                    done: progress_total,
-                    total: progress_total,
-                    phase: SliceProgressPhaseV3::Finalizing,
-                });
-            }
-
-            perf.archive_encode_ns = encode_start.elapsed().as_nanos() as u64;
-            perf.total_ns = total_start.elapsed().as_nanos() as u64;
-            perf.layers = job.total_layers;
-
-            return Ok(perf);
-        }
-    }
-
-    if !requires_png_layers && requires_raw_mask_layers {
-        if let Some(mut stream_encoder) = encoder.create_raw_mask_stream_encoder(job)? {
-            let total_start = std::time::Instant::now();
-            let job_total_layers = job.total_layers;
-            let progress_total = job_total_layers.saturating_add(1);
-            let mut raw_mask_sink = |layer_index: u32, raw_mask: Vec<u8>| {
-                stream_encoder.consume_raw_mask_layer(layer_index, raw_mask)
-            };
-
-            let slicing_progress = on_progress.as_ref().map(|cb| {
-                let cb = cb.clone();
-                Arc::new(move |update: SliceProgressUpdateV3| {
-                    cb(SliceProgressUpdateV3 {
-                        done: update.done.min(job_total_layers),
-                        total: progress_total,
-                        phase: SliceProgressPhaseV3::Slicing,
-                    });
-                }) as ProgressCallbackV3
-            });
-
-            let (_rendered_layers, _layer_area_stats, mut perf) = slice_and_rasterize_v3(
-                job,
-                false,
-                requires_png_layers,
-                false,
-                Some(&mut raw_mask_sink),
-                slicing_progress,
-                cancel_flag,
-            )?;
-
-            if let Some(cb) = on_progress.as_ref() {
-                cb(SliceProgressUpdateV3 {
-                    done: job_total_layers,
-                    total: progress_total,
-                    phase: SliceProgressPhaseV3::Finalizing,
-                });
-            }
-
-            let encode_start = std::time::Instant::now();
-            stream_encoder.finalize_to_path(output_path)?;
-
-            if let Some(cb) = on_progress.as_ref() {
-                cb(SliceProgressUpdateV3 {
-                    done: progress_total,
-                    total: progress_total,
-                    phase: SliceProgressPhaseV3::Finalizing,
-                });
-            }
-
-            perf.archive_encode_ns = encode_start.elapsed().as_nanos() as u64;
-            perf.total_ns = total_start.elapsed().as_nanos() as u64;
-            perf.layers = job.total_layers;
-
-            return Ok(perf);
-        }
-    }
-
-    let total_start = std::time::Instant::now();
-    let (rendered_layers, layer_area_stats, mut perf) = slice_and_rasterize_v3(
-        job,
-        requires_area_stats,
-        requires_png_layers,
-        requires_raw_mask_layers,
-        None,
-        on_progress.clone(),
-        cancel_flag,
-    )?;
-
-    let encode_units = encoder
-        .estimate_encode_progress_units(&rendered_layers)
-        .max(1);
-    let progress_total = job.total_layers.saturating_add(encode_units);
-
-    let encode_progress = on_progress.as_ref().map(|cb| {
-        let cb = cb.clone();
-        move |done: u32, total: u32| {
-            let safe_total = total.max(1);
-            let clamped_done = done.min(safe_total);
-            let normalized =
-                ((clamped_done as u64) * (encode_units as u64) / (safe_total as u64)) as u32;
-            cb(SliceProgressUpdateV3 {
-                done: job.total_layers.saturating_add(normalized),
-                total: progress_total,
-                phase: SliceProgressPhaseV3::Encoding,
-            });
-        }
-    });
-
-    let encode_start = std::time::Instant::now();
-    dispatch_encode_by_format_to_path(
-        job,
-        &rendered_layers,
-        &layer_area_stats,
-        output_path,
-        encode_progress.as_ref().map(|cb| cb as &dyn Fn(u32, u32)),
-    )?;
-
-    if let Some(cb) = on_progress.as_ref() {
-        cb(SliceProgressUpdateV3 {
-            done: progress_total,
-            total: progress_total,
-            phase: SliceProgressPhaseV3::Finalizing,
-        });
-    }
-
-    perf.archive_encode_ns = encode_start.elapsed().as_nanos() as u64;
-    perf.total_ns = total_start.elapsed().as_nanos() as u64;
-    perf.layers = job.total_layers;
-
-    Ok(perf)
-}
-
-impl From<zip::result::ZipError> for SlicerV3Error {
-    fn from(value: zip::result::ZipError) -> Self {
-        Self::Zip(value.to_string())
-    }
-}
-
-impl From<std::io::Error> for SlicerV3Error {
-    fn from(value: std::io::Error) -> Self {
-        Self::Zip(value.to_string())
-    }
-}
-
-impl From<serde_json::Error> for SlicerV3Error {
-    fn from(value: serde_json::Error) -> Self {
-        Self::Json(value.to_string())
-    }
-}
-
-#[allow(dead_code)]
-fn _empty_perf() -> SlicingPerfV3 {
-    SlicingPerfV3::default()
-}
