+++ rust/dragonfruit-slicing-engine/src/metrics.rs
@@ -1,0 +1,25 @@
+//! Performance counters used for diagnostics and UI telemetry.
+
+#[derive(Debug, Clone, Default)]
+pub struct SlicingPerfV3 {
+    pub total_ns: u64,
+    pub index_build_ns: u64,
+    pub render_wall_ns: u64,
+    pub render_ns: u64,
+    pub png_encode_ns: u64,
+    pub archive_encode_ns: u64,
+    pub layers: u32,
+}
+
+impl SlicingPerfV3 {
+    pub fn total_s(&self) -> f64 {
+        self.total_ns as f64 / 1_000_000_000.0
+    }
+
+    pub fn layers_per_second(&self) -> f64 {
+        if self.total_ns == 0 {
+            return 0.0;
+        }
+        (self.layers as f64) / self.total_s().max(1e-9)
+    }
+}
